worker/requirements.txt
fastapi==0.108.0
llama-index==0.9.22
pika==1.3.2
uvicorn==0.25.0
weaviate-client==3.26.0
